fix(type-system): Fix function symbol registration
Parameter symbols and types must be registered in the scope of the
function that we are type checking.
We create a new scope whenever we find a function declaration, but
we were creating the scope way to early. Previously, we were creating the scope
right at the start, before registering the symbol for the function, which would make
the function invisible outside of itself.

Move the scope creation after function registration, and, consequently, move
parameters registration after the function scope is created.