refactor(type-system): Merge symbol table for variables and functions
Previously, for each scope, we had a distinct symbol table for variables and functions.
But currently, there is no reason to define a function and a variable with the same
name in the same scope.

Since scoping fixes shadowing beetween variables and functions too, this is a non-issue.
FunctionType is also a type of Type, so merging these tables is a simplification with no
downsides.